Make it possible to support multiple I2C instances
While attempting to add support for all I2C instance of the STM32L0x2, I
discovered that the macro made it impossible to do that, as that would
have resulted in multiple `impl` blocks for `I2c` with conflicting
method definitions.

I decided to solve that problem in the following way:
- Move the `impl` block for `I2c` out of the macro.
- Abstract over the differences between I2C instances with a new trait,
  `i2c::Instance`.

I believe this approach is superior to the traditional "everything in
the macro" approach, for the following reasons:
- These macros can make the code less readable, as you often have to
  jump between macro definition and macro arguments.
- The code in the macro is deeply indented and the macro can sometimes
  confuse the syntax highlighting, exacerbating the readability
  problems.
- Any compiler error in the macro shows up multiple times, for each
  implementation that the macro generates.
- Multiple nearly identical methods are generated. I haven't checked
  whether this affects the size of the compiled code negatively, but I
  wouldn't be surprised if it did.

For these reasons, I believe that keeping as much code out of the macro
as possible is the better approach.