Remove safe `dma::Transfer` constructors
The problem with these constructors is that they're not flexible enough.
In general, peripheral APIs might want to present DMA transfers to the
user in a different way than they work in hardware.

For example, the AES peripheral supports only transfers of 32-bit words,
but might want to present an API that works with byte slices (I'm
currently working on this). Another example would be SPI, which might
want to hide two simultaneous DMA transfers (for read and write) behind
a single transfer struct.

This commit doesn't make these use cases possible yet, as some
modifications to the unsafe constructor are still required. But it paves
the way, by exposing that constructor to the peripheral APIs.

Please note that none of this exposes any unsafety to the users. They
will still use the same (safe) peripheral APIs to access DMA.