Add `pwm::Timer`
As noted in a previous commit, the current structure of the `pwm` is not
flexible enough to support the capabilities of STM32L0x2 fully.
Specifically the approach of defining the channels in sets that all use
the same alternate function number does not match the hardware.

I believe a better approach is to define each channel separately, as
that provides the most flexibility. I also believe it will be simpler to
make each channel accessible as a separate struct, which requires having
a main struct that has them as fields. This main struct is the new
`pwm::Timer` struct which is added in this commit.

In addition, I've taken the chance to clean up the PWM initialization a
bit:
- I've removed the `PwmExt` struct, replacing it with a constructor.
  This is simple and straight-forward.
- I've replace the old macro with a simpler, more lightweight one. This
  means more code lives outside of the macro, which has a number of
  advantages.