Make type parameter names of `Pwm` match others
I'm a bit torn about this change, because it seems a bit less readable
when looking only at the struct. Unfortunately I can't name them
`Instance` and `Channel`, as that would conflict with the trait names.

However, this makes the struct consistent with most of the rest of the
module, which should improve readability overall. Plus, the
all-uppercase names look weird, as they are usually used for constants
and statics, not types.