Fix RTC prescaler factors
The values used for almost, but not quite, correct. I thought that I'd
set the two prescaler factors to 64 and 512 by writing those values, but
what I actually did was set them to 65 and 513, as per reference manual.

Those new values set them to 128 and 256, which are the values
recommended in a piece of documentation that I've previously missed,
which should set the clock frequency used by the calendar to a clean
1 Hz.