Only disable RTC write protection when needed
I don't see why this would be necessary, but I've been having problem
getting wakeup from Standy mode to work for a while now, and I'm getting
desperate. Only disabling the write protection when requires is like the
documentation and other implementation handle it, so I'd like to be on
the safe side and do that too.