Do not set flash wait states to 1 by default
Back when I implemented this, I was under the assumption that 1 wait
state was a safe default value that should work for all configurations,
although not always be optimal. This turned out to be wrong.

I've seen a weird case, where the USB peripheral could get into a state
where it would no longer send data. This only happened with the number
of wait states set to 1. I don't know how other voltage range and core
frequency configurations might affect this.

I think the best option to now is to leave this configuration up to the
user, which is why I've added a method to the Flash API to configure it.