Changed name of adc payload struct to allow for more states
ADC driver uses a trait parameter to indicate what state the driver is
in. You are only allowed to use functions from the active mode of the
adc if it is in that state.

This patch changes the the name of the Active Struct to a less vague
name, ActiveContinuousDMA. This sets the stage for adding more
operating modes of the ADC.