Added support for adc read_all() dma mode similiar to other drivers
This commit adds a new driver interface for the ADC that mimics the
read_all() and write_all() interfaces for the AES, UART and I2C that
are already present in this crate.

This is because the existing ADC DMA implementation, which allows
iterating over partial DMA results in a continuous manner, contains
some data races, since there is no guarntee the DMA controller is not
writing to the same memory that the user is trying to read.

This new implementation requires the user to move the needed resources
into the driver, and then wait until the dma transfer has
completed. After the DMA transfer has completed, the user can get
back the ADC resources, and do another transfer.

This should also allow the user to do a dma transaction, and then
reconfigure the ADC to sample a different channel.

This implementation is also limited compared to the original: it
only allows sampling a single ADC channel instead of performing a
multiple read of multiple channels. This can be improved in the future
though.

Lastly, the adc_dma.rs example is not quite finished: the current
implementation panics if the buffer size is too big, and it doesnt
provide a good example of a "continuous" adc read, as the DMA is
single buffered instead of double bufferd.