Optimized ADC dma for repeated transfers
This cleans up the ADC DMA oneshot mode to simplify code that executes
multiple transfers back-to-back. Now, you can put the ADC in a "dma
ready" mode, where it is already programmed with the channels to
convert and the proper external trigger, and then you can repeated
call the read_all() function with a buffer that you want to results to
go into.