Added ability to do scan dma on multiple channels
The DMA oneshot implementation now can be set up to scan multiple adc
pins. It still wants to move the resources inside of it during the
conversion, so trait implementations are defined for tuples of adc
inputs so they can be passed in in one group. If this is not
sufficient, the user can implement the trait on his collection type to
specify the flag.