LinkedTimer: Implement initial version
A `LinkedTimer` is initialized with two hardware timers (either
TIM2/TIM3 or TIM21/TIM22). The two timers are configured in master/slave
mode so that an overflow of the master timer triggers an update on the
slave timer. This way, two 16 bit timers can be combined to a single 32
bit timer.