Add a CHANGELOG
Format is loosely based on keepachangelog.com, the format that a lot of
other STM32 HAL crates use as well.

The main issues with single-file changelogs is that they cause a lot of
conflict potential. One solution is writing changelog entries to
separate files, and merging those before a release. However, that
requires additional tooling and is a bit cumbersome.

Instead, I added `CHANGELOG.md merge=union` to `.gitattributes`. I'm not
sure if it always works ideally, but at least it should reduce the
conflict potential a lot. And in the case something breaks, we can
always fix it manually.