Add set_frequency method on PWM channels
This adds a method to PWM channels to change the underlying timer's
frequency. This allows to dynamically change the PWM frequency when
we have associated a channel with a pin. Otherwise this required
unsafe to work around a partial move error because the channel was
moved out of the timer struct.

See #172 for more information