Use either LSI or LSE in RTC (#218)
RTC can be used with an internal or external Low Speed clock (LSE or
LSI).  LSE is more accurate but requires an external crystal on the
board.   This diff adds the possibility to select LSI or LSE when RTC is
created.

Also, modified existing examples to use LSI.  This prevents them from
freezing on boards without an LSE crystal.  None of the examples seem to
rely on high timing accuracy.

Co-authored-by: Javier Cardona <javier.cardona@gmail.com>