Implement ADC self-calibration (#233)
* Implement ADC self-calibration

Implement the ACD self-calibration procedure, which is defined in the
RM377 reference manual as follows:

The ADC has a calibration feature. During the procedure, the ADC
calculates a calibration factor which is internally applied to the ADC
until the next ADC power-off. The application must not use the ADC
during calibration and must wait until it is complete.  Calibration
should be performed before starting A/D convers

* Wait for hardware to clear ADCAL after calibration

The reference manual says that software should wait until ADCAL = 0.
Do that instead of being paranoid and eventually timing out if this does
not happen.

---------

Co-authored-by: Javier Cardona <javier@cardonabits.com>